# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: phpcp Control Panel API
  version: "2.0.0"
  description: |
    API ของ Control Panel สำหรับบริหาร PHP Web Hosting

    **การยืนยันตัวตน** ใช้คุกกี้ session (HttpOnly + SameSite=Strict) ที่เซิร์ฟเวอร์ตั้งให้
    ไม่ใช่ bearer token — จงใจไม่ย้าย token ไปเก็บใน JavaScript เพราะจะเป็นการถอยหลัง
    จากสิ่งที่ระบบมีอยู่แล้ว (PLAN-V2 ข้อ N6)

    **CSRF** ทุกคำขอที่เปลี่ยนข้อมูล (POST/PUT/PATCH/DELETE) ต้องส่ง header
    `X-CSRF-Token` ที่ได้จาก `GET /session` · เซิร์ฟเวอร์ส่ง token ตัวใหม่กลับมาทาง
    response header ชื่อเดียวกันทุกครั้งที่มันเปลี่ยน (ตอนล็อกอิน ยืนยัน 2FA
    หรือตอน session id ถูกหมุนตามรอบ) ฝั่ง client ต้องเก็บค่าล่าสุดไว้ใช้เสมอ
servers:
  - url: https://localhost:8443/api/v2
    description: panel ของเครื่องนี้ (พอร์ตเริ่มต้น 8443)
...
paths:
...
  /sites/{id}/domains:
    parameters:
      - { $ref: "#/components/parameters/SiteId" }
    get:
      tags: [sites]
      summary: โดเมนทั้งหมดของเว็บไซต์นี้
      responses:
        "200":
          description: รายการโดเมน
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/Error" }

    post:
      tags: [sites]
      summary: เพิ่มโดเมนย่อยหรือ alias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [host]
              properties:
                host: { type: string, description: "ชื่อโฮสต์ย่อย หรือโดเมนเต็ม" }
                path: { type: string, description: "โฟลเดอร์ปลายทางของโดเมนย่อย" }
      responses:
        "201":
          description: เพิ่มแล้ว
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }

    put:
      tags: [sites]
      summary: แทนที่รายการโดเมนทั้งชุดตามชนิด
      description: capability คำนวณเองว่าต้องเพิ่มหรือลบอะไรเพื่อให้ตรงกับรายการที่ส่งมา
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domains]
              properties:
                domains:
                  oneOf:
                    - { type: array, items: { type: string } }
                    - { type: string }
                type: { type: string, enum: [alias, subdomain], default: alias }
      responses:
        "200":
          description: ผลการแทนที่
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { type: object }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /sites/{id}/domains/{domain}:
    delete:
      tags: [sites]
      summary: ลบโดเมนย่อยหรือ alias ออกจากเว็บไซต์
      parameters:
        - { $ref: "#/components/parameters/SiteId" }
        - name: domain
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
        "422":
          description: "`VALIDATION_ERROR` — ลบโดเมนหลักแยกไม่ได้"
          $ref: "#/components/responses/Error"
  /domains:
    get:
      tags: [domains]
      summary: โดเมนทั้งหมดที่ผู้เรียกมีสิทธิ์เห็น
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
        - name: q
          in: query
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [primary, subdomain, alias, redirect] }
      responses:
        "200":
          description: รายการโดเมนพร้อมข้อมูลเว็บไซต์ที่สังกัด
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Domain" }
                  meta: { $ref: "#/components/schemas/PageMeta" }
        "401": { $ref: "#/components/responses/Error" }

    post:
      tags: [domains]
      summary: เพิ่มโดเมนโดยระบุเว็บไซต์ปลายทาง
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [site_id, host]
              properties:
                site_id: { type: integer }
                host: { type: string }
                path: { type: string }
      responses:
        "201":
          description: เพิ่มแล้ว
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /domains/{id}:
    delete:
      tags: [domains]
      summary: ลบโดเมน
      parameters:
        - { $ref: "#/components/parameters/DomainId" }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: "`CONFLICT` — เป็นโดเมนหลักของเว็บไซต์ ลบแยกไม่ได้ ต้องลบทั้งเว็บ"
          $ref: "#/components/responses/Error"
...
  /domains/{id}/dns-records:
    parameters:
      - { $ref: "#/components/parameters/DomainId" }
    get:
      tags: [domains]
      summary: DNS record ทั้งหมดของโดเมนนี้
      responses:
        "200":
          description: รายการเรกคอร์ด
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/DnsRecord" }
        "404": { $ref: "#/components/responses/Error" }

    post:
      tags: [domains]
      summary: เพิ่ม DNS record
      description: |
        ค่าถูกตรวจให้ตรงกับชนิดเสมอ — ใส่ IP ลงช่อง CNAME จะถูกปฏิเสธพร้อมบอกทางแก้
        (ให้ใช้ A หรือ AAAA แทน) ซึ่งเป็นความผิดพลาดที่ DNS server จะรับไว้เงียบ ๆ
        แล้วทำให้ชื่อนั้นใช้ไม่ได้ทั้งโดเมน
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, value]
              properties:
                type: { type: string, enum: [A, AAAA, CNAME, MX, TXT, CAA] }
                name: { type: string, default: "@", description: "`@` = โดเมนตัวเอง · `*` = wildcard" }
                value: { type: string, maxLength: 512 }
                ttl: { type: integer, minimum: 60, maximum: 86400, default: 3600 }
                priority:
                  type: integer
                  description: ใช้กับ MX เท่านั้น · ไม่ส่งมาจะได้ค่าปริยาย 10
      responses:
        "201":
          description: |
            เพิ่มแล้ว — `dns_synced`/`dns_message` บอกผลการส่ง zone ไปยัง BIND9 จริง
            (เฟส E3) ล้มได้โดยไม่ทำให้คำขอนี้ล้มเหลวไปด้วย เพราะเรกคอร์ดถูกบันทึกใน
            ฐานข้อมูลสำเร็จแล้วก่อนถึงขั้นนี้ · `dns_synced = false` พร้อม `dns_message`
            อธิบายว่า "ยังไม่ได้เปิด dns.enabled" เป็นผลลัพธ์ปกติของการติดตั้งส่วนใหญ่
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, const: true }
                  data:
                    type: object
                    properties:
                      record_id: { type: integer }
                      domain_id: { type: integer }
                      dns_synced: { type: boolean }
                      dns_message: { type: string }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /dns-records/{id}:
    delete:
      tags: [domains]
      summary: ลบ DNS record
      description: |
        ลบเรกคอร์ดแล้วสั่ง `dns.zone_write` เขียน zone ของโดเมนนั้นใหม่ให้อัตโนมัติ
        (ล้มได้โดยไม่ทำให้การลบล้มเหลว — ดูหมายเหตุที่ `POST /domains/{id}/dns-records`)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: สำเร็จ — คำตอบของคำสั่งพร้อม `actions` ให้หน้าจอทำตาม
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandResponse" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      description: ค่าที่เกินเพดานถูกบีบลงมาที่ 200 ไม่ใช่ปฏิเสธคำขอ
    SiteId:
      name: id
      in: path
      required: true
      schema: { type: integer }
    DomainId:
      name: id
      in: path
      required: true
      schema: { type: integer }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: phpcp_sid
      description: |
        ตั้งโดยเซิร์ฟเวอร์เท่านั้น · HttpOnly + SameSite=Strict + Secure (เมื่อใช้ HTTPS)
        ชื่อจะเป็น `__Host-phpcp_sid` เมื่อเปิด cookie_secure
  schemas:
...
    PageMeta:
      type: object
      required: [page, per_page, total, total_pages]
      properties:
        page: { type: integer }
        per_page: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
...
    Domain:
      type: object
      properties:
        id: { type: integer }
        site_id: { type: integer }
        domain: { type: string }
        type: { type: string, enum: [primary, subdomain, alias, redirect] }
        removable:
          type: boolean
          description: |
            false สำหรับโดเมนหลัก — ส่งมาให้ตรง ๆ เพื่อให้หน้าจอไม่ต้องรู้กฎนี้เอง
            แล้วเผลอแสดงปุ่มลบที่กดแล้วได้ error
        redirect_target: { type: string }
        redirect_code: { type: [integer, "null"] }
        created_at: { type: [integer, "null"] }
        site_name: { type: string, description: "มีเฉพาะตอนดึงจาก /domains" }
        site_domain: { type: string }
        site_status: { type: string }
    DnsRecord:
      type: object
      properties:
        id: { type: integer }
        domain_id: { type: integer }
        type: { type: string, enum: [A, AAAA, CNAME, MX, TXT, CAA] }
        name: { type: string }
        value: { type: string }
        ttl: { type: integer }
        priority:
          type: [integer, "null"]
          description: null สำหรับทุกชนิดยกเว้น MX — "ไม่มีลำดับ" ต่างจาก "ลำดับ 0"
        can_manage:
          type: boolean
          description: สิทธิ์ `domain.manage` ของผู้เรียก — เงื่อนไขปุ่มลบในตารางอ่านได้แค่ค่าในแถวเดียวกัน
...
    CommandResponse:
      type: object
      description: |
        คำตอบของ **คำสั่ง** (POST/PUT/PATCH/DELETE ที่เปลี่ยนสถานะ)

        **ไม่มีคีย์ `data`** — ค่าที่ผู้เรียกต้องใช้อยู่ระดับบนสุดคู่กับ `message`
        และ `actions` · กฎนี้มาจากวิธีที่ Now.js แกะคำตอบ (`response.data.data ??
        response.data`) ซึ่งจะเลือกชั้นในทันทีที่มี `data` ซ้อนอยู่ แล้ว `actions`
        ที่อยู่ชั้นนอกจะไม่ถูกเห็นเลย

        คำตอบสำหรับ**อ่าน** (GET) ยังเป็น `{ok, data, meta}` ตามเดิม เพราะตารางและ
        ฟอร์มผูกกับ `data` โดยตรง
      properties:
        ok: { type: boolean, const: true }
        message: { type: string, description: ข้อความสรุปผลสำหรับคนอ่าน }
        actions: { $ref: "#/components/schemas/ResponseActions" }
      additionalProperties: true
    ResponseActions:
      type: array
      description: |
        คำสั่งหน้าจอที่เซิร์ฟเวอร์ส่งกลับมา — `ResponseHandler` ของ Now.js อ่านคีย์
        `actions` ที่ระดับบนสุดของคำตอบแล้วทำตามลำดับ

        ใช้กับผลลัพธ์ที่ **ต้องให้ผู้ใช้เห็นเนื้อหาในคำตอบ** ไม่ใช่แค่รู้ว่าสำเร็จ
        เช่นรหัสผ่านที่ระบบสุ่มให้ ซึ่งไม่มีที่อื่นให้ดูย้อนหลังอีกเลย

        ชนิดที่ใช้ได้: `notification` `alert` `modal` `redirect` `update` `render`
        `remove` `class` `attribute` `focus` `scroll` `clipboard` `download` `event`
      items:
        type: object
        required: [type]
        properties:
          type: { type: string }
          message: { type: string }
          title: { type: string }
          content: { type: string }
          level: { type: string, enum: [success, info, warning, error] }
          url: { type: string, description: "`reload` = โหลดใหม่ · ที่เหลือคือปลายทางจริง" }
          target: { type: string, description: "ชื่อ data-table ที่จะโหลดใหม่" }
          action: { type: string, enum: [show, open, close, hide] }
...
    Error:
      type: object
      required: [ok, error]
      properties:
        ok: { type: boolean, const: false }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: รหัสตายตัวที่ฝั่งเรียกใช้ตัดสินใจ — ห้ามเทียบจากข้อความ
              enum:
                - VALIDATION_ERROR
                - UNAUTHENTICATED
                - TWO_FACTOR_REQUIRED
                - PASSWORD_CHANGE_REQUIRED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - CSRF_INVALID
                - RATE_LIMITED
                - QUOTA_EXCEEDED
                - PROTECTED_RESOURCE
                - AGENT_UNAVAILABLE
                - EXECUTION_FAILED
                - BAD_REQUEST
                - METHOD_NOT_ALLOWED
                - INTERNAL_ERROR
            message:
              type: string
              description: ข้อความภาษาไทยที่แสดงให้ผู้ใช้เห็นได้ทันที
            fields:
              type: object
              additionalProperties: { type: string }
              description: ชื่อฟิลด์ => เหตุผลที่ค่านั้นไม่ผ่าน (มีเฉพาะ VALIDATION_ERROR)
  responses:
    Error:
      description: ข้อผิดพลาดตามรูปแบบมาตรฐาน
      headers:
        X-CSRF-Token:
          schema: { type: string }
          description: แนบมากับ 419 เสมอ เพื่อให้ลองใหม่ได้ทันทีโดยไม่ต้องรีเฟรชหน้า
        Retry-After:
          schema: { type: integer }
          description: แนบมากับ 429 เสมอ
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
security:
  - sessionCookie: []